Move load fast optimization into _compiler
This should have been there in the first place but I was having some
weird issue with it not being called. Whatever the issue is, it's gone
now.
